💄 별점 컴포넌트 조건부 스타일링 추가
